model_api/PostDescriptionModel/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
mistralai
gtts